Refatoração

+++ src/combobox.cpp
@@ -8,7 +8,7 @@ QVariant ComboBox::getCurrentValue() const { return (currentData()); }
 
 bool ComboBox::setCurrentValue(const QVariant &value) {
   for (int index = 0, size = count(); index < size; ++index) {
-    if (value.toInt() == itemData(index).toInt()) {
+    if (value == itemData(index)) {
       setCurrentIndex(index);
       return true;
     }
